[Critical Fix] DO NOT compile with -Ofast
Verified that the cause of false edge count for raw network is compilation with the -Ofast optimization option.  Do not compile with -Ofast, but instead compile with -O3.  -Ofast causes asynchronous computation which can greatly increase speed, but is not compatible with our variable reference and visibility design.  Work is being done to narrow down the cause of errors that only appear with -Ofast, but NaN MI values are computed for different regulator-target interactions for different subnets.  This could be for a variety of possible reasons.  One reason could be that while the expr_vec_sampled and expr_vec are being loaded into gm and gm_folds[subnet_idx] in MatrixReglistIO.cpp, genes are randomly getting null expression vectors.  This could explain why different subnets suffer from different MI calculations, which pull expression vectors from the genemaps, as does the final MI calculation.  Another possible reason is that vec_x and vec_y are file static variables for the AMPI.cpp class, and there is asynchronous computation with -Ofast compiled.  Work must be done to determine why this error only happens with -Ofast clang compilation, as bugs could 'leak' into other optimization options, such as -O3 which we now use.