CMake: Successfully find OpenMP on x86_64 MacOS
- Critically, include(FindOpenMP) did not always successfully create the
target library for OpenMP_CXX, which was linked to the target executable
ARACNe3_app.  This is because the library installation for OpenMP is not
standardized between homebrew versions for x86 and arm64 Macs.  This
means that ARACNe3 did not successfully build on x86 MacOS, even when
using include(FindOpenMP).  The solution was including a very specific
if condition that paths the x86 libraries for Intel macs and creating an
imported library target.

- ARACNe3_lib does not build because I did not link that target to the
  imported library.  I plan to change the structure of the CMake
dependency so that ARACNe3_lib is built before ARACNe3_app.  This makes
way more sense, as ARACNe3_app can just link to ARACNe3_lib to compile,
without having to double link the imported library.