CMake: Complete overhaul of ARACNe3 targeting
- Changed the targeting scheme so that ARACNe3_lib was built first,
  which builds the functions into a usable library, mainly for testing.
If only to use for testing, it could have had PUBLIC includes for the
ARACNe3 headers and PRIVATE includes for the OMP and BOOST headers, but
everything must be public because when ARACNe3_app links ARACNe3_lib, it
must also see all the includes for its own source code compilation. Note
the oddity of the fact that a main() function is included in the static
library, yet this is "fixed", so to say, at runtime (but it is bad
practice).  The purpose of creating ARACNe3 as a library is primarily
for testing and for creating ARACNe3_app, so this is fine.  Note that
the linked libraries must also be PUBLIC so that ARACNe3_app can link
them when linking to the lib.