CMake: Rm main() from libARACNe3_lib.a
- Target ARACNe3_lib did not need main(), so ARACNe3.cpp was removed from SOURCES and added into the add_executable() source files list

Note about commit signing: The verification is only that the global
git.config e-mail matches the e-mail in the gpg key.  If this is the
case, then the commit is verified.  The actual account associated with
pushing the changes is related to the account that you add the rsa ssh
key for cloning/pushing/pulling.  Hence, with the account, you can make
verified commits from a totally different e-mail.